requirements.txt
playwright==1.41.1
pytest==8.0.0
pytest-playwright==0.4.4
pytest-xdist==3.5.0
allure-pytest==2.13.2
pyyaml==6.0.1
python-dotenv==1.0.0
